viejo/requirements.txt
asttokens==2.4.1
attrs==24.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.53.1
imutils==0.5.4
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mpmath==1.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.0.1
opencv-contrib-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==26.1.0
referencing==0.35.1
rpds-py==0.20.0
scipy==1.14.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.2
tenacity==9.0.0
tk==0.1.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
wcwidth==0.2.13
